+++ fsm/common/src/commonMain/kotlin/dev/inmo/micro_utils/fsm/common/StatesHandlerHolder.kt
@@ -3,36 +3,66 @@ package dev.inmo.micro_utils.fsm.common
 import kotlin.reflect.KClass
 
 /**
+ * Define checkable holder which can be used to precheck that this handler may handle incoming [State]
+ */
+interface CheckableHandlerHolder<I : State, O : State> : StatesHandler<I, O> {
+    suspend fun checkHandleable(state: O): Boolean
+}
+
+/**
  * Default realization of [StatesHandler]. It will incapsulate checking of [State] type in [checkHandleable] and class
  * casting in [handleState]
  */
-class StatesHandlerHolder<I : State>(
-    private val inputKlass: KClass<I>,
-    private val strict: Boolean = false,
-    private val delegateTo: StatesHandler<I>
-) : StatesHandler<State> {
+class CustomizableHandlerHolder<I : O, O : State>(
+    private val delegateTo: StatesHandler<I, O>,
+    private val filter: suspend (state: O) -> Boolean
+) : CheckableHandlerHolder<I, O> {
     /**
      * Checks that [state] can be handled by [delegateTo]. Under the hood it will check exact equality of [state]
      * [KClass] and use [KClass.isInstance] of [inputKlass] if [strict] == false
      */
-    fun checkHandleable(state: State) = state::class == inputKlass || (!strict && inputKlass.isInstance(state))
+    override suspend fun checkHandleable(state: O) = filter(state)
 
     /**
      * Calls [delegateTo] method [StatesHandler.handleState] with [state] casted to [I]. Use [checkHandleable]
      * to be sure that this [StatesHandlerHolder] will be able to handle [state]
      */
-    override suspend fun StatesMachine.handleState(state: State): State? {
-        return delegateTo.run { handleState(state as I) }
+    override suspend fun StatesMachine<in O>.handleState(state: I): O? {
+        return delegateTo.run { handleState(state) }
     }
 }
 
-@Deprecated("Renamed", ReplaceWith("StatesHandlerHolder"))
-typealias StateHandlerHolder<T> = StatesHandlerHolder<T>
+fun <I : O, O : State> StateHandlerHolder(
+    inputKlass: KClass<I>,
+    strict: Boolean = false,
+    delegateTo: StatesHandler<I, O>
+) = CustomizableHandlerHolder(
+    StatesHandler<O, O> {
+        delegateTo.run { handleState(it as I) }
+    },
+    if (strict) {
+        { it::class == inputKlass }
+    } else {
+        { inputKlass.isInstance(it) }
+    }
+)
 
-inline fun <reified T : State> StatesHandler<T>.holder(
+inline fun <reified I : O, O : State> StateHandlerHolder(
+    strict: Boolean = false,
+    delegateTo: StatesHandler<I, O>
+) = StateHandlerHolder(I::class, strict, delegateTo)
+
+inline fun <reified I : O, O: State> StatesHandler<I, O>.holder(
     strict: Boolean = true
-) = StatesHandlerHolder(
-    T::class,
+) = StateHandlerHolder<I, O>(
+    I::class,
     strict,
     this
+)
+
+inline fun <I : O, O: State> StatesHandler<I, O>.holder(
+    noinline filter: suspend (state: State) -> Boolean
+) = CustomizableHandlerHolder<O, O>(
+    { this@holder.run { handleState(it as I) } },
+    filter
 )