+++ fsm/common/src/commonMain/kotlin/dev/inmo/micro_utils/fsm/common/dsl/FSMBuilder.kt
@@ -4,36 +4,47 @@ import dev.inmo.micro_utils.fsm.common.*
 import dev.inmo.micro_utils.fsm.common.managers.*
 import kotlin.reflect.KClass
 
-class FSMBuilder(
-    var statesManager: StatesManager = DefaultStatesManager(InMemoryDefaultStatesManagerRepo()),
-    var defaultStateHandler: StatesHandler<State>? = StatesHandler { null }
+class FSMBuilder<T : State>(
+    var statesManager: StatesManager<T> = DefaultStatesManager(InMemoryDefaultStatesManagerRepo()),
+    var defaultStateHandler: StatesHandler<T, T>? = StatesHandler { null }
 ) {
-    private var states = mutableListOf<StatesHandlerHolder<*>>()
+    private var states = mutableListOf<CustomizableHandlerHolder<T, T>>()
 
-    fun <I : State> add(kClass: KClass<I>, handler: StatesHandler<I>) {
-        states.add(StatesHandlerHolder(kClass, false, handler))
+    fun <I : T> add(kClass: KClass<I>, handler: StatesHandler<I, T>) {
+        states.add(StateHandlerHolder(kClass, false, handler))
     }
 
-    fun <I : State> addStrict(kClass: KClass<I>, handler: StatesHandler<I>) {
-        states.add(StatesHandlerHolder(kClass, true, handler))
+    fun <I : T> add(filter: suspend (state: State) -> Boolean, handler: StatesHandler<I, T>) {
+        states.add(handler.holder(filter))
+    }
+
+    fun <I : T> addStrict(kClass: KClass<I>, handler: StatesHandler<I, T>) {
+        states.add(StateHandlerHolder(kClass, true, handler))
+    }
+
+    inline fun <reified I : T> onStateOrSubstate(handler: StatesHandler<I, T>) {
+        add(I::class, handler)
+    }
+
+    inline fun <reified I : T> strictlyOn(handler: StatesHandler<I, T>) {
+        addStrict(I::class, handler)
+    }
+
+    inline fun <reified I : T> doWhen(
+        noinline filter: suspend (state: State) -> Boolean,
+        handler: StatesHandler<I, T>
+    ) {
+        add(filter, handler)
     }
 
     fun build() = StatesMachine(
         statesManager,
         states.toList().let { list ->
-            defaultStateHandler ?.let { list + it.holder(false) } ?: list
+            defaultStateHandler ?.let { list + it.holder { true } } ?: list
         }
     )
 }
 
-inline fun <reified I : State> FSMBuilder.onStateOrSubstate(handler: StatesHandler<I>) {
-    add(I::class, handler)
-}
-
-inline fun <reified I : State> FSMBuilder.strictlyOn(handler: StatesHandler<I>) {
-    addStrict(I::class, handler)
-}
-
-fun buildFSM(
-    block: FSMBuilder.() -> Unit
-): StatesMachine = FSMBuilder().apply(block).build()
+fun <T : State> buildFSM(
+    block: FSMBuilder<T>.() -> Unit
+): StatesMachine<T> = FSMBuilder<T>().apply(block).build()