+++ fsm/common/src/commonMain/kotlin/dev/inmo/micro_utils/fsm/common/managers/DefaultStatesManager.kt
@@ -9,25 +9,25 @@ import kotlinx.coroutines.sync.withLock
 /**
  * Implement this repo if you want to use some custom repo for [DefaultStatesManager]
  */
-interface DefaultStatesManagerRepo {
+interface DefaultStatesManagerRepo<T : State> {
     /**
      * Must save [state] as current state of chain with [State.context] of [state]
      */
-    suspend fun set(state: State)
+    suspend fun set(state: T)
     /**
      * Remove exactly [state]. In case if internally [State.context] is busy with different [State], that [State] should
      * NOT be removed
      */
-    suspend fun removeState(state: State)
+    suspend fun removeState(state: T)
     /**
      * @return Current list of available and saved states
      */
-    suspend fun getStates(): List<State>
+    suspend fun getStates(): List<T>
 
     /**
      * @return Current state by [context]
      */
-    suspend fun getContextState(context: Any): State?
+    suspend fun getContextState(context: Any): T?
 
     /**
      * @return Current state by [context]
@@ -43,21 +43,21 @@ interface DefaultStatesManagerRepo {
  * key. In case when this callback will returns true, the state placed on [State.context] of new will be replaced by
  * new state by using [endChain] with that state
  */
-class DefaultStatesManager(
-    private val repo: DefaultStatesManagerRepo = InMemoryDefaultStatesManagerRepo(),
-    private val onContextsConflictResolver: suspend (old: State, new: State, currentNew: State) -> Boolean = { _, _, _ -> true }
-) : StatesManager {
-    private val _onChainStateUpdated = MutableSharedFlow<Pair<State, State>>(0)
-    override val onChainStateUpdated: Flow<Pair<State, State>> = _onChainStateUpdated.asSharedFlow()
-    private val _onStartChain = MutableSharedFlow<State>(0)
-    override val onStartChain: Flow<State> = _onStartChain.asSharedFlow()
-    private val _onEndChain = MutableSharedFlow<State>(0)
-    override val onEndChain: Flow<State> = _onEndChain.asSharedFlow()
+class DefaultStatesManager<T : State>(
+    private val repo: DefaultStatesManagerRepo<T> = InMemoryDefaultStatesManagerRepo(),
+    private val onContextsConflictResolver: suspend (old: T, new: T, currentNew: T) -> Boolean = { _, _, _ -> true }
+) : StatesManager<T> {
+    private val _onChainStateUpdated = MutableSharedFlow<Pair<T, T>>(0)
+    override val onChainStateUpdated: Flow<Pair<T, T>> = _onChainStateUpdated.asSharedFlow()
+    private val _onStartChain = MutableSharedFlow<T>(0)
+    override val onStartChain: Flow<T> = _onStartChain.asSharedFlow()
+    private val _onEndChain = MutableSharedFlow<T>(0)
+    override val onEndChain: Flow<T> = _onEndChain.asSharedFlow()
 
     private val mapMutex = Mutex()
 
-    override suspend fun update(old: State, new: State) = mapMutex.withLock {
-        val stateByOldContext: State? = repo.getContextState(old.context)
+    override suspend fun update(old: T, new: T) = mapMutex.withLock {
+        val stateByOldContext: T? = repo.getContextState(old.context)
         when {
             stateByOldContext != old -> return@withLock
             stateByOldContext == null || old.context == new.context -> {
@@ -76,26 +76,26 @@ class DefaultStatesManager(
         }
     }
 
-    override suspend fun startChain(state: State) = mapMutex.withLock {
+    override suspend fun startChain(state: T) = mapMutex.withLock {
         if (!repo.contains(state.context)) {
             repo.set(state)
             _onStartChain.emit(state)
         }
     }
 
-    private suspend fun endChainWithoutLock(state: State) {
+    private suspend fun endChainWithoutLock(state: T) {
         if (repo.getContextState(state.context) == state) {
             repo.removeState(state)
             _onEndChain.emit(state)
         }
     }
 
-    override suspend fun endChain(state: State) {
+    override suspend fun endChain(state: T) {
         mapMutex.withLock {
             endChainWithoutLock(state)
         }
     }
 
-    override suspend fun getActiveStates(): List<State> = repo.getStates()
+    override suspend fun getActiveStates(): List<T> = repo.getStates()
 
 }